+++ src/controllers/index.ts
@@ -1,3 +1,3 @@
 export { default as RateController } from './RateController';
-export { default as subscribe } from './subscribe';
-export { default as unsubscribe } from './unsubscribe';
+export { default as SubscribeController } from './SubscribeController';
+export { default as UnsubscribeController } from './UnsubscribeController';
