--- src/controllers/subscribe.ts
@@ -1,40 +1,0 @@
-import { Request, Response } from 'express';
-
-import responseMessages from '../constants/responseMessages';
-import EmailSubscriptionService from '../services/EmailSubscriptionService';
-
-const { EMAIL_ALREADY_EXISTS, INVALID_PAYLOAD } = responseMessages;
-
-const subscribe = async (req: Request, res: Response): Promise<Response> => {
-  const { email: requestEmail } = req.body as { email: string };
-
-  const emailSubscription = new EmailSubscriptionService(requestEmail);
-
-  try {
-    const validationError = emailSubscription.validate();
-    if (validationError) {
-      return res.status(INVALID_PAYLOAD.code).send({
-        validationError,
-      });
-    }
-    const alreadyExists = emailSubscription.isExists();
-
-    if (alreadyExists) {
-      res.status(EMAIL_ALREADY_EXISTS.code).send({
-        error: EMAIL_ALREADY_EXISTS.error,
-      });
-    } else {
-      await emailSubscription.subscribe();
-
-      res.send(emailSubscription.email);
-    }
-  } catch (error) {
-    console.error(error);
-
-    return res
-      .status(INVALID_PAYLOAD.code)
-      .send({ error: INVALID_PAYLOAD.error });
-  }
-};
-
-export default subscribe;
